# Support JDK 23 in optimizer (via upgrade to patched ASM 9.7.0)