--- a/src/compiler/scala/tools/nsc/backend/jvm/analysis/BackendUtils.scala
+++ b/src/compiler/scala/tools/nsc/backend/jvm/analysis/BackendUtils.scala
@@ -86,6 +86,7 @@ abstract class BackendUtils extends PerRunInit {
     case "20" => asm.Opcodes.V20
     case "21" => asm.Opcodes.V21
     case "22" => asm.Opcodes.V22
+    case "23" => asm.Opcodes.V23
     // to be continued...
   })