--- a/versions.properties
+++ b/versions.properties
@@ -6,7 +6,7 @@ starr.version=2.13.14
 #  - scala-compiler: jline (% "optional")
 # Other usages:
 #  - scala-asm: jar content included in scala-compiler
-scala-asm.version=9.6.0-scala-1
+scala-asm.version=9.7.0-scala-2
 
 # REPL
 jline.version=3.26.3