looks like a safe upgrade for 2.13.14, knock on wood: https://asm.ow2.io/versions.html

@Philippus I did this myself (using your #10576 as a guide) because 2.13.14 is imminent, but I leave it to you to do Scala 2.12 and Scala 3, if you're available