@SethTisue 2024-04-09
```
[error] 	scala.tools.nsc.reporters.WConfTest
[error] 	scala.tools.nsc.backend.jvm.OptimizedBytecodeTest
[error] 	scala.tools.nsc.backend.jvm.opt.MethodLevelOptsTest
```