@Philippus 2024-04-09
Maybe it's a Jenkins thing? https://scala-ci.typesafe.com/job/scala-2.12.x-validate-main/5414/console says "No space left on device".