@SethTisue 2024-04-09
```
% cat S.scala
class C {
  def t(as: Listt): Unit = {
    map(as, (_: Any) => return)
  }
  final def map(x: Listt, f: Any => Any): Any = {
    if (x eq Nill) "" else f("")
  }
}
object Nill extends Listt
class Listt
% ~/scala.213/build/pack/bin/scalac '-opt:inline:**' S.scala
error: Error while emitting C
-1 is out of bounds (min 0, max 0)
1 error
% ~/scala.213/build/pack/bin/scalac S.scala
```