@lrytz 2024-04-18
Performance looks fine, I compared the same compiler with asm 9.6 (60f4d19) and 9.7 (bf0dd33)

```
# scalap corpus

# old asm
[info] HotScalacBenchmark.compile                                         latest  -opt:inline:**       false  2.13.14-bin-60f4d19-SNAPSHOT    scalap  sample  427  732,721 ± 6,250  ms/op
# new asm
[info] HotScalacBenchmark.compile                                         latest  -opt:inline:**       false  2.13.14-bin-bf0dd33-SNAPSHOT    scalap  sample  418  739,404 ± 5,886  ms/op
# without optimizer
[info] HotScalacBenchmark.compile                                         latest                    false  2.13.14-bin-bf0dd33-SNAPSHOT    scalap  sample  802  381,399 ± 3,115  ms/op

# scala corpus

#old asm
[info] HotScalacBenchmark.compile                                         latest  -opt:inline:**       false  2.13.14-bin-60f4d19-SNAPSHOT     scala  sample   30  46607,106 ± 677,909  ms/op
# new asm
[info] HotScalacBenchmark.compile                                         latest  -opt:inline:**       false  2.13.14-bin-bf0dd33-SNAPSHOT     scala  sample   30  46911,333 ± 593,243  ms/op
# without optimizer
[info] HotScalacBenchmark.compile                                         latest                    false  2.13.14-bin-bf0dd33-SNAPSHOT     scala  sample   30  20785,852 ± 504,137  ms/op
```

My first fix in this PR (in `newExceptionValue`) seems fine. For the fix I did in 9.7.0-scala-2 (https://github.com/scala/scala-asm/commit/cb8082501d8600bb58a5d321ee4e45319190f2e3), let's wait and see what the maintainers say.